made all updates conrad asked for
